initial projects list from github api
