+++ src/components/projects.rs
@@ -1,43 +1,16 @@
 use crate::components::card::Card;
 use log::info;
-use reqwasm::http::Request;
-use serde::{Deserialize, Serialize};
 use sycamore::prelude::*;
 
-// API that counts visits to the web-page
-const API_BASE_URL: &str = "https://api.github.com/users/benjaminboruff/repos";
-
-#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq, Eq, Hash)]
-pub struct Repository {
-    name: String,
-    html_url: String,
-    description: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Default, Debug, Clone)]
-#[serde(transparent)]
-struct Repositories {
-    repositories: Vec<Repository>,
-}
-
-async fn fetch_all_projects<G: Html>() -> Result<Repositories, reqwasm::Error> {
-    let url = format!("{}", API_BASE_URL);
-    let resp = Request::get(&url).send().await?;
-
-    let resp_str = resp.text().await?;
-    let repositories: Repositories = serde_json::from_str(&resp_str.as_str())?;
-    Ok(repositories)
-}
+use crate::Repository;
 
 #[component]
-pub async fn Projects<'cx, G: Html>(cx: Scope<'cx>) -> View<G> {
-    let github = fetch_all_projects::<G>().await.unwrap_or_default();
-    let repos = github.repositories;
-    info!("{:?}", repos);
-    let projects = create_signal(cx, repos);
+pub fn Projects<G: Html>(cx: Scope) -> View<G> {
+    let repos_data: &Signal<Vec<Repository>> = use_context(cx);
+    info!("{:?}", repos_data);
     view! {cx,
         Keyed(
-            iterable=projects,
+            iterable=repos_data,
             view=|cx, repo| view! { cx,
                 Card(repo=repo){}
             },
